Jaetaan web security config kahteen tiedostoon.
