+++ src/main/java/fi/vm/sade/varda/rekisterointi/configuration/VirkailijaWebSecurityConfiguration.java
@@ -1,0 +1,99 @@
+package fi.vm.sade.varda.rekisterointi.configuration;
+
+import fi.vm.sade.java_utils.security.OpintopolkuCasAuthenticationFilter;
+import fi.vm.sade.javautils.kayttooikeusclient.OphUserDetailsServiceImpl;
+import fi.vm.sade.properties.OphProperties;
+import fi.vm.sade.varda.rekisterointi.util.Constants;
+import org.jasig.cas.client.validation.Cas20ProxyTicketValidator;
+import org.jasig.cas.client.validation.TicketValidator;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.cas.ServiceProperties;
+import org.springframework.security.cas.authentication.CasAuthenticationProvider;
+import org.springframework.security.cas.web.CasAuthenticationEntryPoint;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
+
+import javax.servlet.Filter;
+
+@Configuration
+@Order(1)
+@EnableWebSecurity
+public class VirkailijaWebSecurityConfiguration extends WebSecurityConfigurerAdapter {
+
+    private static final String VIRKAILIJA_ROLE = "APP_YKSITYISTEN_REKISTEROITYMINEN_CRUD";
+    private static final String VIRKAILIJA_PATH_CLOB = "/virkailija/**";
+
+    private final OphProperties ophProperties;
+
+    VirkailijaWebSecurityConfiguration(OphProperties ophProperties) {
+        this.ophProperties = ophProperties;
+    }
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.headers().disable().csrf().disable();
+        http.authorizeRequests()
+                .antMatchers(VIRKAILIJA_PATH_CLOB).hasRole(VIRKAILIJA_ROLE)
+                .and()
+                .addFilterBefore(virkailijaAuthenticationProcessingFilter(), BasicAuthenticationFilter.class)
+                .exceptionHandling()
+                .authenticationEntryPoint(virkailijaAuthenticationEntryPoint());
+    }
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) {
+        auth.authenticationProvider(virkailijaAuthenticationProvider());
+    }
+
+    @Bean
+    public ServiceProperties serviceProperties() {
+        ServiceProperties serviceProperties = new ServiceProperties();
+        serviceProperties.setService(ophProperties.getProperty("varda-rekisterointi.virkailija") + "/j_spring_cas_security_check");
+        serviceProperties.setSendRenew(false);
+        serviceProperties.setAuthenticateAllArtifacts(true);
+        return serviceProperties;
+    }
+
+    @Bean
+    public Filter virkailijaAuthenticationProcessingFilter() throws Exception {
+        OpintopolkuCasAuthenticationFilter casAuthenticationFilter = new OpintopolkuCasAuthenticationFilter(serviceProperties());
+        casAuthenticationFilter.setAuthenticationManager(authenticationManager());
+        casAuthenticationFilter.setFilterProcessesUrl("/virkailija/j_spring_cas_security_check");
+        return casAuthenticationFilter;
+    }
+
+    @Bean
+    public AuthenticationEntryPoint virkailijaAuthenticationEntryPoint() {
+        CasAuthenticationEntryPoint casAuthenticationEntryPoint = new CasAuthenticationEntryPoint();
+        casAuthenticationEntryPoint.setLoginUrl(ophProperties.url("cas.login"));
+        casAuthenticationEntryPoint.setServiceProperties(serviceProperties());
+        return casAuthenticationEntryPoint;
+    }
+
+    @Bean
+    public AuthenticationProvider virkailijaAuthenticationProvider() {
+        CasAuthenticationProvider casAuthenticationProvider = new CasAuthenticationProvider();
+        String host = ophProperties.url("kayttooikeus-service.host");
+        casAuthenticationProvider.setUserDetailsService(new OphUserDetailsServiceImpl(host, Constants.CALLER_ID));
+        casAuthenticationProvider.setServiceProperties(serviceProperties());
+        casAuthenticationProvider.setTicketValidator(ticketValidator());
+        casAuthenticationProvider.setKey("varda-rekisterointi");
+        return casAuthenticationProvider;
+    }
+
+    @Bean
+    public TicketValidator ticketValidator() {
+        Cas20ProxyTicketValidator ticketValidator = new Cas20ProxyTicketValidator(ophProperties.url("cas.base"));
+        ticketValidator.setAcceptAnyProxy(true);
+        return ticketValidator;
+    }
+
+
+}
